rename HIRT_Multichannel_Convolution/{ConvolveMultichannel => multichannel}.hpp
