+++ include/HIRT_Multichannel_Convolution/multichannel.hpp
